function reorg
